Auth implementado

+++ app/src/main/java/pe/com/valegrei/carwashapp/network/request/ReqCambiarClave.kt
@@ -1,3 +1,3 @@
 package pe.com.valegrei.carwashapp.network.request
 
-class ReqCambiarClave(var correo: String, var clave: String, var codigo: Long)
+class ReqCambiarClave(var correo: String, var clave: String, var codigo: Int)
